cloud-functions/search/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.14.0
google-cloud-aiplatform>=1.38.0
flask==3.0.0
